validando o delete das rotinas e dos exercicios

+++ src/controllers/exercises_controllers.js
@@ -8,10 +8,15 @@ import { getSeriesById } from "../models/exercises_models.js";
 export async function createExerciseController(req, res){
     try{
         const id = req.params.id;
-        const {exercise_name} = req.body;
+        const data = req.body;
         
-        const create = await createExercise({ routine_id: id , exercise_name });
-        res.status(200).json({message: "sucess in creating", id: create});
+        delete data.id;
+        delete data.routine_id;
+
+        if(data.exercise_name){
+            const create = await createExercise({ id , exercise_name: data.exercise_name });
+            res.status(200).json({message: "sucess in creating", id: create});
+        }
     }catch(error){
         console.error(error.message)
         res.status(400).json({message: "Failure to create"})
@@ -22,10 +27,15 @@ export async function createExerciseController(req, res){
 export async function editExerciseController (req, res){
     try{
         const id = req.params.id;
-        const exercise_name = req.body;
+        const data = req.body;
 
-        const edit = await editExercise({exercise_name, exercise_id: id});
-        res.status(200).json({message: "Success when updating"});
+        delete data.id;
+        delete data.routine_id;
+
+        if(data.exercise_name){
+            const edit = await editExercise({exercise_name, exercise_id: id});
+            res.status(200).json({message: "Success when updating"});
+        }
     }catch(error){
         console.error(error.message);
         res.status(400).json({message: "Failure to update"});
@@ -36,9 +46,23 @@ export async function editExerciseController (req, res){
 export async function deleteExerciseController(req, res){
     try{
         const id = req.params.id;
+        const {confirm} = req.body;
+        const regex = /^Deletar$/;
 
-        const delExercise = await deleteExercise({id})
-        res.status(200).json({message: "Exercise deleted!"})
+        if(regex.test(String(confirm))){
+            const delExercise = await deleteExercise({id})
+           
+
+            if(delExercise){
+                res.status(200).json({message: "Exercise deleted"})
+            }else{
+                res.status(400).json({message: "Exercise not founded"})
+            }
+
+        }else{
+            res.status(400).json({message: "Confirmation necessary"})
+        }
+
     }catch(error){
         console.error(error.message);
         res.status(400).json({message: " failed to delete "})
